+++ src/main/java/com/github/standobyte/jojo/mixin/client/controls/MouseHandlerInvoker.java
@@ -1,6 +1,7 @@
 package com.github.standobyte.jojo.mixin.client.controls;
 
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
 import org.spongepowered.asm.mixin.gen.Invoker;
 
 import net.minecraft.client.MouseHandler;
@@ -10,4 +11,10 @@ public interface MouseHandlerInvoker {
 
 	@Invoker("onMove")
 	void jojo_ripples$onMove(long window, double x, double y);
+
+	@Accessor("xpos")
+	void jojo_ripples$setXpos(double x);
+
+	@Accessor("ypos")
+	void jojo_ripples$setYpos(double y);
 }